+++ NEWS.md
@@ -1,10 +1,16 @@
+
+# eventglm 1.4.0
+
+New features: 
+
+- New pseudo module for infinitesimal jackknife that allows for delayed entry/left truncation.
+
 # eventglm 1.3.0
 
 New features: 
 
 - Allow id argument in cumincglm and rmeanglm. When id is present, geese.fit will be used instead of glm.fit, with clusters indicated by the id. This is to account for clustered data, for example.
 - Informative error message if there is missing data in the censoring model
-- New pseudo module for infinitesimal jackknife that allows for delayed entry/left truncation.
 
 # eventglm 1.2.2